+++ openmeetings-web/src/main/java/org/apache/openmeetings/web/user/profile/MessagesContactsPanel.java
@@ -255,7 +255,7 @@ public class MessagesContactsPanel extends UserBasePanel {
 				unreadMessages.clear();
 				String sort = getSort() == null ? "" : "m." + getSort().getProperty();
 				boolean isAsc = getSort() == null ? true : getSort().isAscending();
-				return getDao().get(getUserId(), selectedFolderModel.getObject(), search, sort, isAsc, (int)first, (int)count).iterator();
+				return getDao().get(getUserId(), selectedFolderModel.getObject(), search, sort, isAsc, first, count).iterator();
 			}
 
 			@Override
@@ -418,7 +418,7 @@ public class MessagesContactsPanel extends UserBasePanel {
 
 			@Override
 			public Iterator<? extends UserContact> iterator(long first, long count) {
-				return getBean(UserContactDao.class).get(getUserId(), (int)first, (int)count).iterator();
+				return getBean(UserContactDao.class).get(getUserId(), first, count).iterator();
 			}
 
 			@Override
